+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- fj-doc-mod-fop, recover Apache FOP version <https://github.com/fugerit-org/fj-doc/issues/571>
+
 ## [8.17.4] - 2025-11-04
 
 ### Changed